+++ PowerShellGet/PSModule.psm1
@@ -137,9 +137,10 @@ $script:NuGetProvider = $null
 # PowerShellGetFormatVersion is in the form of Major.Minor.  
 # Minor is incremented for the backward compatible format change.
 # Major is incremented for the breaking change.
-$script:CurrentPSGetFormatVersion = "1.0"
+$script:PSGetRequireLicenseAcceptanceFormatVersion = [Version]'2.0'
+$script:CurrentPSGetFormatVersion = $script:PSGetRequireLicenseAcceptanceFormatVersion
 $script:PSGetFormatVersion = "PowerShellGetFormatVersion"
-$script:SupportedPSGetFormatVersionMajors = @("1")
+$script:SupportedPSGetFormatVersionMajors = @("1","2")
 $script:ModuleReferences = 'Module References'
 $script:AllVersions = "AllVersions"
 $script:Filter      = "Filter"
@@ -846,7 +847,7 @@ function Publish-Module
         $Credential,
 
         [Parameter()] 
-        [ValidateSet("1.0")]
+        [ValidateSet("2.0")]
         [Version]
         $FormatVersion,
 
@@ -1164,13 +1165,14 @@ function Publish-Module
         $tempModulePath = Microsoft.PowerShell.Management\Join-Path -Path $script:TempPath `
                               -ChildPath "$(Microsoft.PowerShell.Utility\Get-Random)\$moduleName"
 
-        if(-not $FormatVersion)
-        {
-            $tempModulePathForFormatVersion = $tempModulePath
-        }
-        elseif ($FormatVersion -eq "1.0")
+
+        if ($FormatVersion -eq "1.0")
         {
             $tempModulePathForFormatVersion = Microsoft.PowerShell.Management\Join-Path $tempModulePath "Content\Deployment\$script:ModuleReferences\$moduleName"
+        }
+        else
+        {
+            $tempModulePathForFormatVersion = $tempModulePath
         }
 
         $null = Microsoft.PowerShell.Management\New-Item -Path $tempModulePathForFormatVersion -ItemType Directory -Force -ErrorAction SilentlyContinue -WarningAction SilentlyContinue -Confirm:$false -WhatIf:$false
@@ -1569,7 +1571,11 @@ function Save-Module
 
         [Parameter()]
         [switch]
-        $Force
+        $Force,
+
+        [Parameter()]
+        [switch]
+        $AcceptLicense
     )
 
     Begin
@@ -1798,7 +1804,11 @@ function Install-Module
 
         [Parameter()]
         [switch]
-        $Force
+        $Force,
+
+        [Parameter()]
+        [switch]
+        $AcceptLicense
     )
 
     Begin
@@ -2016,7 +2026,11 @@ function Update-Module
 
         [Parameter()]
         [Switch]
-        $Force
+        $Force,
+
+        [Parameter()]
+        [switch]
+        $AcceptLicense
     )
 
     Begin
@@ -3238,7 +3252,11 @@ function Save-Script
 
         [Parameter()]
         [switch]
-        $Force
+        $Force,
+
+        [Parameter()]
+        [switch]
+        $AcceptLicense
     )
 
     Begin
@@ -3467,7 +3485,11 @@ function Install-Script
 
         [Parameter()]
         [switch]
-        $Force
+        $Force,
+
+        [Parameter()]
+        [switch]
+        $AcceptLicense
     )
 
     Begin
@@ -3749,7 +3771,11 @@ function Update-Script
 
         [Parameter()]
         [Switch]
-        $Force
+        $Force,
+
+        [Parameter()]
+        [switch]
+        $AcceptLicense
     )
 
     Begin
@@ -7795,29 +7821,60 @@ function ValidateAndGet-ScriptDependencies
                                         WarningAction = 'SilentlyContinue'
                                         Debug = $DebugPreference
                                     }
+            $ReqScriptInfo = @{}
+
             if ($PSBoundParameters.ContainsKey('Credential'))
             {
                 $FindScriptArguments.Add('Credential',$Credential)
             }
-
-            if($DependentScriptInfo.ExternalScriptDependencies -contains $requiredScript)
+            
+            if (-not ($requiredScript -match '^(?<ScriptName>[^:]+)(:(?<Version>[^:\s]+))?$'))
             {
-                Write-Verbose -Message ($LocalizedData.SkippedScriptDependency -f $requiredScript)
+                $message = $LocalizedData.FailedToParseRequiredScripts -f ($requiredScript)
+
+                ThrowError `
+                    -ExceptionName "System.ArgumentException" `
+                    -ExceptionMessage $message `
+                    -ErrorId "FailedToParseRequiredScripts" `
+                    -CallerPSCmdlet $CallerPSCmdlet `
+                    -ErrorCategory InvalidOperation
+            }
+
+            $scriptName = $Matches['ScriptName']
+            if ($DependentScriptInfo.ExternalScriptDependencies -contains $scriptName)
+            {
+                Write-Verbose -Message ($LocalizedData.SkippedScriptDependency -f $scriptName)
 
                 continue
             }
 
-            $FindScriptArguments['Name'] = $requiredScript
-            $ReqScriptInfo = @{}
-            $ReqScriptInfo['Name'] = $requiredScript
+            if ($Matches.Keys -Contains 'Version')
+            {
+                $ReqScriptInfo = ValidateAndGet-NuspecVersionString -Version $Matches['Version']
+                
+                if($ReqScriptInfo.Keys -Contains 'RequiredVersion')
+                {
+                    $FindScriptArguments['RequiredVersion'] = $ReqScriptInfo['RequiredVersion']
+                }
+                elseif($ReqScriptInfo.Keys -Contains 'MinimumVersion')
+                {
+                    $FindScriptArguments['MinimumVersion'] = $ReqScriptInfo['MinimumVersion']
+                }
+                if($ReqScriptInfo.Keys -Contains 'MaximumVersion')
+                {
+                    $FindScriptArguments['MaximumVersion'] = $ReqScriptInfo['MaximumVersion']
+                }
+            }
 
+            $ReqScriptInfo['Name'] = $scriptName
+            $FindScriptArguments['Name'] = $scriptName
             $psgetItemInfo = Find-Script @FindScriptArguments  | 
-                                        Microsoft.PowerShell.Core\Where-Object {$_.Name -eq $requiredScript} | 
+                                        Microsoft.PowerShell.Core\Where-Object {$_.Name -eq $scriptName} | 
                                             Microsoft.PowerShell.Utility\Select-Object -Last 1 -ErrorAction Ignore
 
             if(-not $psgetItemInfo)
             {
-                $message = $LocalizedData.UnableToResolveScriptDependency -f ('script', $requiredScript, $DependentScriptInfo.Name, $Repository, 'ExternalScriptDependencies')
+                $message = $LocalizedData.UnableToResolveScriptDependency -f ('script', $scriptName, $DependentScriptInfo.Name, $Repository, 'ExternalScriptDependencies')
                 ThrowError -ExceptionName "System.InvalidOperationException" `
                             -ExceptionMessage $message `
                             -ErrorId "UnableToResolveScriptDependency" `
@@ -7830,6 +7887,146 @@ function ValidateAndGet-ScriptDependencies
     }
 
     return $DependenciesDetails
+}
+
+function ValidateAndGet-NuspecVersionString
+{
+    param(
+        [Parameter(Mandatory=$true)]
+        [string]
+        $Version
+    )
+
+    $versionPattern = '^((?<MinRule>[\[\(])?((?<MinVersion>[^:\(\[\)\]\,]+))?((?<Comma>[\,])?(?<MaxVersion>[^:\(\[\)\]\,]+)?)?(?<MaxRule>[\]\)])?)$'
+    $VersionInfo = @{}
+
+    if ( -not ($Version -match $versionPattern))
+    {
+        $message = $LocalizedData.FailedToParseRequiredScriptsVersion -f ('Invalid Version format', $Version, $LocalizedData.RequiredScriptVersoinFormat)
+        Write-Verbose $message
+        ThrowError -ExceptionName "System.ArgumentException" `
+                    -ExceptionMessage $message `
+                    -ErrorId "UnableToResolveScriptDependency" `
+                    -CallerPSCmdlet $CallerPSCmdlet `
+                    -ErrorCategory InvalidOperation
+    }
+
+    if ($Matches.Keys -Contains 'MinRule' -xor $Matches.Keys -Contains 'MaxRule')
+    {
+        $message = $LocalizedData.FailedToParseRequiredScriptsVersion -f ('Minimum and Maximum inclusive/exclusive condition mismatch', $Version, $LocalizedData.RequiredScriptVersoinFormat)
+        Write-Verbose $message
+        ThrowError -ExceptionName "System.ArgumentException" `
+                    -ExceptionMessage $message `
+                    -ErrorId "UnableToResolveScriptDependency" `
+                    -CallerPSCmdlet $CallerPSCmdlet `
+                    -ErrorCategory InvalidOperation
+    }
+
+    if (-not ($Matches.Keys -Contains 'MinVersion' -or $Matches.Keys -Contains 'MaxVersion'))
+    {
+        $message = $LocalizedData.FailedToParseRequiredScriptsVersion -f ('No version.', $Version, $LocalizedData.RequiredScriptVersoinFormat)
+        Write-Verbose $message
+        ThrowError -ExceptionName "System.ArgumentException" `
+                    -ExceptionMessage $message `
+                    -ErrorId "UnableToResolveScriptDependency" `
+                    -CallerPSCmdlet $CallerPSCmdlet `
+                    -ErrorCategory InvalidOperation
+    }
+
+    if ((-not ($Matches.Keys -Contains 'MinRule' -and $Matches.Keys -Contains 'MaxRule')) -and $Matches.Keys -Contains 'Comma')
+    {
+        $message = $LocalizedData.FailedToParseRequiredScriptsVersion -f ('Invalid version format', $Version, $LocalizedData.RequiredScriptVersoinFormat)
+        Write-Verbose $message
+        ThrowError -ExceptionName "System.ArgumentException" `
+                    -ExceptionMessage $message `
+                    -ErrorId "UnableToResolveScriptDependency" `
+                    -CallerPSCmdlet $CallerPSCmdlet `
+                    -ErrorCategory InvalidOperation
+    }
+
+    if ($Matches.Keys -Contains 'MaxRule' -and -not ($Matches['MaxRule'] -eq ']') )
+    {
+        $message = $LocalizedData.FailedToParseRequiredScriptsVersion -f ('Maximum version condition should be inclusive', $Version, $LocalizedData.RequiredScriptVersoinFormat)
+        Write-Verbose $message
+        ThrowError -ExceptionName "System.ArgumentException" `
+                    -ExceptionMessage $message `
+                    -ErrorId "UnableToResolveScriptDependency" `
+                    -CallerPSCmdlet $CallerPSCmdlet `
+                    -ErrorCategory InvalidOperation
+    }
+
+    if ($Matches.Keys -Contains 'MinVersion' -and $Matches.Keys -Contains 'MaxVersion')
+    {
+        if ($Matches.Keys -Contains 'MinRule' -and $Matches.Keys -Contains 'MaxRule')
+        {
+            if ($Matches['MinRule'] -eq '[')
+            {
+                $VersionInfo['MinimumVersion'] = $Matches['MinVersion']
+                $VersionInfo['MaximumVersion'] = $Matches['MaxVersion']
+            }
+            else
+            {
+                $message = $LocalizedData.FailedToParseRequiredScriptsVersion -f ('Minimum version condition should be inclusive', $Version, $LocalizedData.RequiredScriptVersoinFormat)
+                Write-Verbose $message
+                ThrowError -ExceptionName "System.ArgumentException" `
+                            -ExceptionMessage $message `
+                            -ErrorId "UnableToResolveScriptDependency" `
+                            -CallerPSCmdlet $CallerPSCmdlet `
+                            -ErrorCategory InvalidOperation
+            }
+        }
+        else
+        {
+            $message = $LocalizedData.FailedToParseRequiredScriptsVersion -f ('Minimum and Maximum inclusive/exclusive condition mismatch', $Version, $LocalizedData.RequiredScriptVersoinFormat)
+            Write-Verbose $message
+            ThrowError -ExceptionName "System.ArgumentException" `
+                        -ExceptionMessage $message `
+                        -ErrorId "UnableToResolveScriptDependency" `
+                        -CallerPSCmdlet $CallerPSCmdlet `
+                        -ErrorCategory InvalidOperation
+        }
+
+        return $VersionInfo
+    }
+
+    if ($Matches.Keys -Contains 'MinVersion')
+    {
+        if ($Matches.Keys -Contains 'MinRule' -and $Matches.Keys -Contains 'MaxRule')
+        {
+            if (($Matches['MinRule'] -eq '[') -and ($Matches['MaxRule'] -eq ']'))
+            {
+                $VersionInfo['RequiredVersion'] = $Matches['MinVersion']
+                return $VersionInfo
+            }
+        }
+        else
+        {
+            $VersionInfo['MinimumVersion'] = $Matches['MinVersion']
+            return $VersionInfo
+        }
+
+        $message = $LocalizedData.FailedToParseRequiredScriptsVersion -f ("Minimum and Maximum version rules should be inclusive for 'RequiredVersion'", $Version, $LocalizedData.RequiredScriptVersoinFormat)
+        Write-Verbose $message
+        ThrowError -ExceptionName "System.ArgumentException" `
+                    -ExceptionMessage $message `
+                    -ErrorId "UnableToResolveScriptDependency" `
+                    -CallerPSCmdlet $CallerPSCmdlet `
+                    -ErrorCategory InvalidOperation
+    }
+
+    if ($Matches.Keys -Contains 'MaxVersion')
+    {
+        $VersionInfo['MaximumVersion'] = $Matches['MaxVersion']
+        return $VersionInfo
+    }
+
+    $message = $LocalizedData.FailedToParseRequiredScriptsVersion -f ("Failed to parse version string", $Version, $LocalizedData.RequiredScriptVersoinFormat)
+    Write-Verbose $message
+    ThrowError -ExceptionName "System.ArgumentException" `
+                -ExceptionMessage $message `
+                -ErrorId "UnableToResolveScriptDependency" `
+                -CallerPSCmdlet $CallerPSCmdlet `
+                -ErrorCategory InvalidOperation
 }
 
 function ValidateAndGet-RequiredModuleDetails
@@ -8211,6 +8408,7 @@ function Publish-PSArtifactUtility
     $Author = $null
     $CompanyName = $null
     $Copyright = $null
+    $requireLicenseAcceptance = "false"
 
     if($PSModuleInfo)
     {
@@ -8250,6 +8448,64 @@ function Publish-PSArtifactUtility
             if( -not $ProjectUri -and $PSModuleInfo.PrivateData.PSData["ProjectUri"])
             { 
                 $ProjectUri = $PSModuleInfo.PrivateData.PSData.ProjectUri
+            }
+            if($PSModuleInfo.PrivateData.PSData["RequireLicenseAcceptance"])
+            {
+                $requireLicenseAcceptance = $PSModuleInfo.PrivateData.PSData.requireLicenseAcceptance.ToString().ToLower()
+                if($requireLicenseAcceptance -eq "true")
+                {
+                    if($FormatVersion -and ($FormatVersion.Major -lt $script:PSGetRequireLicenseAcceptanceFormatVersion.Major))
+                    {
+                        $message = $LocalizedData.requireLicenseAcceptanceNotSupported -f($FormatVersion)
+                        ThrowError -ExceptionName "System.InvalidOperationException" `
+                        -ExceptionMessage $message `
+                        -ErrorId "requireLicenseAcceptanceNotSupported" `
+                        -CallerPSCmdlet $PSCmdlet `
+                        -ErrorCategory InvalidData
+                    }
+
+                    if(-not $LicenseUri)
+                    {
+                        $message = $LocalizedData.LicenseUriNotSpecified
+                        ThrowError -ExceptionName "System.InvalidOperationException" `
+                            -ExceptionMessage $message `
+                            -ErrorId "LicenseUriNotSpecified" `
+                            -CallerPSCmdlet $PSCmdlet `
+                            -ErrorCategory InvalidData
+                    }
+
+                    $licenseFile = Get-ChildItem -path $NugetPackageRoot -filter License.txt
+                    if(-not $licenseFile)
+                    {
+                        $message = $LocalizedData.LicenseTxtNotFound
+                        ThrowError -ExceptionName "System.InvalidOperationException" `
+                        -ExceptionMessage $message `
+                        -ErrorId "LicenseTxtNotFound" `
+                        -CallerPSCmdlet $PSCmdlet `
+                        -ErrorCategory InvalidData
+                    }
+
+                    if((Get-Content $licenseFile.FullName) -eq $null)
+                    {
+                        $message = $LocalizedData.LicenseTxtEmpty
+                        ThrowError -ExceptionName "System.InvalidOperationException" `
+                        -ExceptionMessage $message `
+                        -ErrorId "LicenseTxtEmpty" `
+                        -CallerPSCmdlet $PSCmdlet `
+                        -ErrorCategory InvalidData
+                    }
+
+                    #RequireLicenseAcceptance is true, License uri and license.txt exist. Bump Up the FormatVersion
+                    if(-not $FormatVersion)
+                    {
+                        $FormatVersion = $script:CurrentPSGetFormatVersion
+                    }
+                }
+                elseif($requireLicenseAcceptance -ne "false")
+                {
+                    $InvalidValueForRequireLicenseAcceptance = $LocalizedData.InvalidValueBoolean -f ($requireLicenseAcceptance, "requireLicenseAcceptance")
+                    Write-Warning -Message $InvalidValueForRequireLicenseAcceptance
+                }                
             }
         }
     }
@@ -8451,7 +8707,14 @@ function Publish-PSArtifactUtility
             $VersionString = "$($Dependency.MinimumVersion)"
         }
 
-        $dependencies += "<dependency id='$($ModuleName)' version='$($VersionString)' />"
+        if ([System.string]::IsNullOrWhiteSpace($VersionString))
+        {
+            $dependencies += "<dependency id='$($ModuleName)'/>"
+        }
+        else 
+        {
+            $dependencies += "<dependency id='$($ModuleName)' version='$($VersionString)' />"
+        }
     }
     
     # Populate the nuspec elements
@@ -8465,11 +8728,11 @@ function Publish-PSArtifactUtility
         <owners>$(Get-EscapedString -ElementValue "$CompanyName")</owners>
         <description>$(Get-EscapedString -ElementValue "$Description")</description>
         <releaseNotes>$(Get-EscapedString -ElementValue "$ReleaseNotes")</releaseNotes>
+        <requireLicenseAcceptance>$($requireLicenseAcceptance.ToString())</requireLicenseAcceptance>
         <copyright>$(Get-EscapedString -ElementValue "$Copyright")</copyright>
         <tags>$(if($Tags){ Get-EscapedString -ElementValue ($Tags -join ' ')})</tags>
         $(if($LicenseUri){
-        "<licenseUrl>$(Get-EscapedString -ElementValue "$LicenseUri")</licenseUrl>
-        <requireLicenseAcceptance>true</requireLicenseAcceptance>"
+         "<licenseUrl>$(Get-EscapedString -ElementValue "$LicenseUri")</licenseUrl>"
         })
         $(if($ProjectUri){
         "<projectUrl>$(Get-EscapedString -ElementValue "$ProjectUri")</projectUrl>"
@@ -8744,7 +9007,7 @@ function ValidateAndAdd-PSScriptInfoEntry
 
         $script:RequiredScripts { 
                                     $KeyName = $script:RequiredScripts
-                                    $Value = $Value -split '[,\s+]' | Microsoft.PowerShell.Core\Where-Object {$_}
+                                    $Value = $Value -split ',(?=[^\[^\(]\w(?!\w+[\)\]]))|\s' | Microsoft.PowerShell.Core\Where-Object {$_}
                                 }
 
         $script:ExternalScriptDependencies { 
@@ -8915,6 +9178,7 @@ function Get-DynamicOptions
                     Write-Output -InputObject (New-DynamicOption -Category $category -Name DscResource -ExpectedType StringArray -IsRequired $false)
                     Write-Output -InputObject (New-DynamicOption -Category $category -Name RoleCapability -ExpectedType StringArray -IsRequired $false)
                     Write-Output -InputObject (New-DynamicOption -Category $category -Name Command -ExpectedType StringArray -IsRequired $false)
+                    Write-Output -InputObject (New-DynamicOption -Category $category -Name 'AcceptLicense' -ExpectedType Switch -IsRequired $false)
                 }
 
         Source  {
@@ -10389,6 +10653,7 @@ function Install-PackageUtility
     $IsSavePackage = $false
     $Scope = $null
     $NoPathUpdate = $false
+    $AcceptLicense = $false
 
     # take the fastPackageReference and get the package object again.
     $parts = $fastPackageReference -Split '[|]'
@@ -10519,6 +10784,23 @@ function Install-PackageUtility
                     elseif($Force -eq 'true')
                     {
                         $Force = $true
+                    }
+                }
+            }
+
+            if($options.ContainsKey('AcceptLicense'))
+            {
+                $AcceptLicense = $options['AcceptLicense']
+
+                if($AcceptLicense.GetType().ToString() -eq 'System.String')
+                {
+                    if($AcceptLicense -eq 'false')
+                    {
+                        $AcceptLicense = $false
+                    }
+                    elseif($AcceptLicense -eq 'true')
+                    {
+                        $AcceptLicense = $true
                     }
                 }
             }
@@ -10798,6 +11080,9 @@ function Install-PackageUtility
 
             $installedPkgs = $provider.InstallPackage($script:FastPackRefHashtable[$fastPackageReference], $newRequest)
 
+            $YesToAll = $false
+            $NoToAll = $false
+
             foreach($pkg in $installedPkgs)
             {
                 if($request.IsCanceled)
@@ -10896,14 +11181,63 @@ function Install-PackageUtility
                         Write-Error -Message $message -ErrorId "NotSupportedPowerShellGetFormatVersion" -Category InvalidOperation
                         continue
                     }
-                
-                    if(-not $psgItemInfo.PowerShellGetFormatVersion)
+
+                    if($psgItemInfo.PowerShellGetFormatVersion -eq "1.0")
                     {
-                        $sourceModulePath = Microsoft.PowerShell.Management\Join-Path $tempDestination $pkg.Name
+                        $sourceModulePath = Microsoft.PowerShell.Management\Join-Path $tempDestination "$($pkg.Name)\Content\*\$script:ModuleReferences\$($pkg.Name)"
                     }
                     else
                     {
-                        $sourceModulePath = Microsoft.PowerShell.Management\Join-Path $tempDestination "$($pkg.Name)\Content\*\$script:ModuleReferences\$($pkg.Name)"
+                        $sourceModulePath = Microsoft.PowerShell.Management\Join-Path $tempDestination $pkg.Name
+                    }
+                    
+                    #Prompt if module requires license Acceptance
+                    $requireLicenseAcceptance = $false
+                    if($psgItemInfo.PowerShellGetFormatVersion -and
+                       $psgItemInfo.PowerShellGetFormatVersion -ge $script:PSGetRequireLicenseAcceptanceFormatVersion)
+                     {
+                        if($psgItemInfo.AdditionalMetadata -and $psgItemInfo.AdditionalMetadata.requireLicenseAcceptance)
+                        {
+                              $requireLicenseAcceptance = $psgItemInfo.AdditionalMetadata.requireLicenseAcceptance
+                        }
+                    }
+
+                    if($requireLicenseAcceptance -eq $true)
+                    {
+                        if($Force -and -not($AcceptLicense))
+                        {
+                            $message = $LocalizedData.ForceAcceptLicense -f $pkg.Name
+
+                            ThrowError -ExceptionName "System.ArgumentException" `
+                                       -ExceptionMessage $message `
+                                       -ErrorId "ForceAcceptLicense" `
+                                       -CallerPSCmdlet $PSCmdlet `
+                                       -ErrorCategory InvalidArgument
+                        }
+
+                        If (-not ($YesToAll -or $NoToAll -or $AcceptLicense))
+                        {
+                            if(-not(Test-Path -path "$sourceModulePath\License.txt" -PathType Leaf))
+                            {
+                                $message = $LocalizedData.LicenseTxtNotFound
+
+                                ThrowError -ExceptionName "System.ArgumentException" `
+                                           -ExceptionMessage $message `
+                                           -ErrorId "LicenseTxtNotFound" `
+                                           -CallerPSCmdlet $PSCmdlet `
+                                           -ErrorCategory ObjectNotFound
+                            }                            
+                            $LicenseFilePath = Join-Path $sourceModulePath "License.txt"
+                            $FormattedEula = (Get-Content -Path $LicenseFilePath) -Join "`r`n"
+                            $message = $FormattedEula + "`r`n" + ($LocalizedData.AcceptanceLicenseQuery -f $pkg.Name)
+                            $title = $LocalizedData.AcceptLicense
+                            $result = $request.ShouldContinue($message, $title, [ref]$yesToAll, [ref]$NoToAll)
+                            if(($result -eq $false) -or ($NoToAll -eq $true))
+                            {
+                                Write-Warning -Message $LocalizedData.UserDeclinedLicenseAcceptance
+                                return
+                            }
+                        }
                     }
 
                     $CurrentModuleInfo = $null
@@ -13007,7 +13341,13 @@ function Update-ModuleManifest
         [Parameter()]
         [ValidateNotNullOrEmpty()]
         [String[]]
-        $PackageManagementProviders
+        $PackageManagementProviders,
+
+        [Parameter()]
+        [switch]
+        $RequireLicenseAcceptance
+
+
     )
 
     if(-not (Microsoft.PowerShell.Management\Test-Path -Path $Path -PathType Leaf))
@@ -13482,6 +13822,10 @@ function Update-ModuleManifest
         {
             $Data["IconUri"] = $IconUri
         }
+        if($RequireLicenseAcceptance)
+        {
+            $Data["RequireLicenseAcceptance"] = $RequireLicenseAcceptance
+        }
 
         if($ReleaseNotes)
         {
@@ -13704,6 +14048,9 @@ function Get-PrivateData
         # ReleaseNotes of this module
         # ReleaseNotes = ''
 
+        # Flag to indicate whether the module requires explicit user acceptance for install/update/save
+        # RequireLicenseAcceptance = $false
+
         # External dependent modules of this module
         # ExternalModuleDependencies = ''
 
@@ -13721,18 +14068,33 @@ function Get-PrivateData
     $IconUri = $PrivateData["IconUri"] | %{"'$_'"}
     $ReleaseNotesEscape = $PrivateData["ReleaseNotes"] -Replace "'","''"
     $ReleaseNotes = $ReleaseNotesEscape | %{"'$_'"}
-    $ExternalModuleDependencies = $PrivateData["ExternalModuleDependencies"] -join "','" | %{"'$_'"} 
-    
-    $DefaultProperties = @("Tags","LicenseUri","ProjectUri","IconUri","ReleaseNotes","ExternalModuleDependencies")
+    $ExternalModuleDependencies = $PrivateData["ExternalModuleDependencies"] -join "','" | %{"'$_'"}     
+    $RequireLicenseAcceptance = $PrivateData["RequireLicenseAcceptance"]
+    $DefaultProperties = @("Tags","LicenseUri","ProjectUri","IconUri","ReleaseNotes","ExternalModuleDependencies","RequireLicenseAcceptance")
 
     $ExtraProperties = @()
     foreach($key in $PrivateData.Keys)
     {
         if($DefaultProperties -notcontains $key)
-        {
+        {            
             $PropertyString = "#"+"$key"+ " of this module"
             $PropertyString += "`r`n    "
-            $PropertyString += $key +" = " + "'"+$PrivateData[$key]+"'"
+            if(($PrivateData[$key]).GetType().IsArray)
+            { 
+                $PropertyString += $key +" = " +" @("               
+                $PrivateData[$key] | % { $PropertyString += "'" + $_ +"'" + "," }
+                if($PrivateData[$key].Length -ge 1) 
+                {
+                    #Remove extra ,
+                    $PropertyString = $PropertyString -Replace ".$"
+                }
+                $PropertyString += ")"
+            }
+            else
+            {
+                $PropertyString += $key +" = " + "'"+$PrivateData[$key]+"'"
+            }            
+
             $ExtraProperties += ,$PropertyString
         }
     }
@@ -13777,6 +14139,12 @@ function Get-PrivateData
     {
         $ReleaseNotesLine = "ReleaseNotes = "+$ReleaseNotes
     }
+    $RequireLicenseAcceptanceLine = "# RequireLicenseAcceptance = `$false"
+    if($RequireLicenseAcceptance)
+    {
+        $RequireLicenseAcceptanceLine = "RequireLicenseAcceptance = `$true"
+    }
+
     $ExternalModuleDependenciesLine ="# ExternalModuleDependencies = @()"
     if($ExternalModuleDependencies -ne "''")
     {
@@ -13804,6 +14172,9 @@ function Get-PrivateData
 
         # ReleaseNotes of this module
         $ReleaseNotesLine
+
+        # Flag to indicate whether the module requires explicit user acceptance for install/update/save
+        $RequireLicenseAcceptanceLine
 
         # External dependent modules of this module
         $ExternalModuleDependenciesLine
@@ -13834,6 +14205,9 @@ function Get-PrivateData
 
         # ReleaseNotes of this module
         $ReleaseNotesLine
+
+        # Flag to indicate whether the module requires explicit user acceptance for install/update
+        $RequireLicenseAcceptanceLine
 
         # External dependent modules of this module
         $ExternalModuleDependenciesLine