+++ Tests/PSGetPublishModule.Tests.ps1
@@ -644,7 +644,7 @@ Describe PowerShell.PSGet.PublishModuleTests -Tags 'BVT','InnerLoop' {
 
         New-ModuleManifest -Path "$ModuleBase\$ModuleName.psd1" -ModuleVersion $version -Description $Description -Author $Author -CompanyName $CompanyName -Copyright $CopyRight
 
-        Publish-Module -Path $ModuleBase -NuGetApiKey $script:ApiKey -ReleaseNotes $ReleaseNotes -Tags $Tags -LicenseUri $LicenseUri -ProjectUri $ProjectUri -FormatVersion "1.0" -WarningAction SilentlyContinue
+        Publish-Module -Path $ModuleBase -NuGetApiKey $script:ApiKey -ReleaseNotes $ReleaseNotes -Tags $Tags -LicenseUri $LicenseUri -ProjectUri $ProjectUri -WarningAction SilentlyContinue
 
         RemoveItem -path $ModuleBase
 
@@ -660,8 +660,7 @@ Describe PowerShell.PSGet.PublishModuleTests -Tags 'BVT','InnerLoop' {
         AssertEqualsCaseInsensitive $psgetItemInfo.CopyRight $CopyRight "CopyRight should be same as the published one"
         Assert       ($psgetItemInfo.Tags -contains $($Tags[0])) "Tags ($($psgetItemInfo.Tags)) should contain the published one ($($Tags[0]))"
         Assert       ($psgetItemInfo.Tags -contains $($Tags[1])) "Tags ($($psgetItemInfo.Tags)) should contain the published one ($($Tags[1]))"
-        AssertEqualsCaseInsensitive $psgetItemInfo.LicenseUri $LicenseUri "LicenseUri should be same as the published one"
-        AssertEqualsCaseInsensitive $psgetItemInfo.PowerShellGetFormatVersion ([Version]$script:CurrentPSGetFormatVersion) "PowerShellGetFormatVersion $($psgetItemInfo.PowerShellGetFormatVersion) should match the current PowerShellGetFormatVersion $script:CurrentPSGetFormatVersion"
+        AssertEqualsCaseInsensitive $psgetItemInfo.LicenseUri $LicenseUri "LicenseUri should be same as the published one"        
     }
 
     # Purpose: Test Publish-Module cmdlet gets the PSData properties from the module manifest file