+++ Tests/PSGetUpdateModuleManifest.Tests.ps1
@@ -98,7 +98,18 @@ Describe PowerShell.PSGet.UpdateModuleManifest -Tags 'BVT','InnerLoop' {
         AssertEquals $newModuleInfo.FileList[1] $oldModuleInfo.FileList[1] "FileList[0] should be $($oldModuleInfo.FileList[1])"
         #Make sure the additioanl properties inside PrivateData remain the same
         AssertEquals $newModuleInfo.PrivateData.PackageManagementProviders $oldModuleInfo.PrivateData.PackageManagementProviders "PackageManagement Providers should be $($oldModuleInfo.PrivateData.PackageManagementProviders)"
-        AssertEquals $newModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions $oldModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions "SupportedPowerShellGetFormatVersions should be $($oldModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions)"
+        if($newModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions -is [Array])
+        {
+            AssertEquals $newModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions.Count $oldModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions.Count "SupportedPowerShellGetFormatVersions count should be $($oldModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions.Count)"
+            foreach($ver in $oldModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions)
+            {                
+	            Assert ($newModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions -contains ($ver)) "SupportedPowerShellGetFormatVersions should contain $($ver)"
+            }
+        }
+        else
+        {
+            AssertEquals $newModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions $oldModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions "SupportedPowerShellGetFormatVersions should be $($oldModuleInfo.PrivateData.SupportedPowerShellGetFormatVersions)"
+        }
     } `
     -Skip:$($IsWindows -eq $False)