+++ CHANGELOG.md
@@ -1,10 +1,25 @@
 # Changelog
 
+## 1.5.0.0
+
+New features
+* Added support for modules requiring license acceptance (#150)
+  - [Documentation](https://docs.microsoft.com/en-us/powershell/gallery/psget/module/RequireLicenseAcceptance)
+
+* Added version for REQUIREDSCRIPTS (#162)
+  - Enabled following scenarios for REQUIREDSCRIPTS
+    - [1.0] - RequiredVersion
+    - [1.0,2.0] - Min and Max Version
+    - (,1.0] - Max Version
+    - 1.0 - Min Version
+
+Bug fixes
+* Fixed empty version value in nuspec (#157)
+
 ## 1.1.3.2
 * Disabled PowerShellGet Telemetry on PS Core as PowerShell Telemetry APIs got removed in PowerShell Core beta builds. (#153)
 * Fixed for DateTime format serialization issue. (#141)
 * Update-ModuleManifest should add ExternalModuleDependencies value as a collection. (#129)
-
 
 ## 1.1.3.1